Fixed some bugs in the object copier.  One prevented it from copying some deep objects. It can now handle copying objects from multiple databases into one database.
Abstracted the code in OTViewer which works with user data.  It now uses a OTUserSession object
to work with the user data.  This interface has saving and opening methods as well as asking
if things have changed.

More methods in OTViewer are public now code using it can do more with it.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk@13521 6e01202a-0783-4428-890a-84243c50cc2b