Add option to show user data warning
This warning used to only be enabled if the code was in single user mode, but didn't have any data to be saved.  It was easiest safest to just allow this warning to be explicitly set.  There are enough options that figuring out when it should be shown is not simple.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk@24643 6e01202a-0783-4428-890a-84243c50cc2b