Reafactor otrunk to track authored and non-authored databases separately.
This fixes an inadvertent problem where getReferences() was returning object ids from the user session database, but not composite/transient ids. When the objects were then loaded, it was nulling out all values that were set in the base database but not set in the overlay database. getReferences() now only works with "authored" databases -- the original database otrunk was initialized with and any databases that it loads (and they load and so forth) via includes and whatnot.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk@25623 6e01202a-0783-4428-890a-84243c50cc2b